Move CA1708 to ApiDesignGuidelines analyzers - no functional change.
