--- src/Microsoft.ApiDesignGuidelines.Analyzers/UnitTests/IdentifiersShouldDifferByMoreThanCaseTests.Fixer.cs
@@ -1,33 +1,0 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Test.Utilities;
-using Microsoft.CodeAnalysis.UnitTests;
-using Xunit;
-
-namespace Microsoft.ApiDesignGuidelines.Analyzers.UnitTests
-{
-    public class IdentifiersShouldDifferByMoreThanCaseFixerTests : CodeFixTestBase
-    {
-        protected override DiagnosticAnalyzer GetBasicDiagnosticAnalyzer()
-        {
-            return new BasicIdentifiersShouldDifferByMoreThanCaseAnalyzer();
-        }
-
-        protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
-        {
-            return new CSharpIdentifiersShouldDifferByMoreThanCaseAnalyzer();
-        }
-
-        protected override CodeFixProvider GetBasicCodeFixProvider()
-        {
-            return new BasicIdentifiersShouldDifferByMoreThanCaseFixer();
-        }
-
-        protected override CodeFixProvider GetCSharpCodeFixProvider()
-        {
-            return new CSharpIdentifiersShouldDifferByMoreThanCaseFixer();
-        }
-    }
-}
